Create sighand.c
This program can receive and handle three different signals in both busy wait and blocking wait. The signals that this program can handle are: 35, 36 and 37. It exits if the signal 37 was received. To execute this program, just compile it and call it passing as argument 0 if you'd like busy wait or 1 if you'd like blocking wait.