Create producer-consumer_pipes.c
This program implements a producer-consumer using pipes as communication mechanism. To execute this program, just compile it and call it passing as argument the number of random and always increasing numbers that the producer should generates. These numbers are sent to consumer and the consumer checks if they are prime or not. When all numbers are generated, the producer send the number zero that indicates to consumer that it should finish.