Create producer.c
This program implements the producer component of the program Producer-Consumer. It creates a TCP socket and binds it to the port 4000. After this, the socket starts to listen the port until the consumer component connects to it. When it occurs, the producer begins the production of N random and always increasing numbers, as specified by the argument passed during the program calling. For each number generated, the producer waits for a response of consumer component, that is, it waits for the prime test result. After the production of all numbers, the producer sends the number zero for consumer indicating that it should finish.