Create consumer.c
This program implements the consumer component of the "Producer-Consumer". It creates a socket and connects itself with the producer socket. After this, it waits for a number sent from producer component. When this occur, the consumer checks if this number is or not prime. The result is sent for producer. When the received number is 0, the consumer finalizes its execution.